feat: make simplepush protocol optional
Preliminary work for removing simplepush protocol by making it optional
(default on)

BREAKING: Simplepush may now be disabled on the server at any time in
the future, and may occur without warning.

Issue #799
